Add Searchform component
