+++ frontend/13f-app/src/components/MainSection.js
@@ -1,13 +1,8 @@
 import React from 'react';
 import '../App.css';
-import { Button } from './Button';
-import SearchBox from './SearchBox';
+import '../App.css';
+import { SearchForm } from './SearchForm.js';
 import './MainSection.css';
-
-const divStyle = { float: 'left', padding: '20px', margin: '20px'};
-const date = new Date();
-const today = date.toISOString().substr(0, 10);
-
 
 function MainSection() {
     return (
@@ -15,12 +10,8 @@ function MainSection() {
             <div className='main-container'>
                 <h1>COMPANY FINANCIAL HOLDINGS</h1>
                 <p>Search 13F Filings by Company Name and Date</p>
-                <SearchBox />
-                <div style={divStyle}>
-                <input type='date' className='date' defaultValue={today}></input>
-                <input type='date' className='date' defaultValue={today}></input>
-                </div>
-                <Button buttonStyle='btn--outline' buttonSize='btn--large'>SEARCH</Button>     
+                <SearchForm />
+                     
             </div>
                      
         </>
